use size_t for memory buffer sizes

+++ parse.h
@@ -10,8 +10,8 @@ struct parser_output {
 	char *output;
 
 	char *input;
-	int len;
-	int pos;
+	size_t len;
+	size_t pos;
 };
 
 #endif
